Avoid signed overflow reading a big-endian uint32
get_bigendian_uint32_t shifted base[0] left by 24 without casting first. A
uint8_t promotes to int, so any byte >= 0x80 made that shift signed overflow -
undefined behaviour - in the header and map parser, on data that comes straight
from the file.

Real files never hit it: every field read through this helper holds either a
small count or a four-character codec tag, and those tags are ASCII, so the
high byte is always below 0x80. A malformed file only has to set one high bit.
Found by fuzzing mutated headers under UBSan:

  libchdr_chd.c:763:18: runtime error: left shift of 228 by 24 places
  cannot be represented in type 'int'
    #0 get_bigendian_uint32_t
    #1 header_read
    #2 chd_open_core_file_callbacks

get_bigendian_uint48 and get_bigendian_uint64_t already cast for exactly this
reason; this one was inconsistent with them.

Output is unchanged for well-formed files. Verified against 561 malformed CHDs
generated from all 17 corpus seed codecs - truncations and header and body
mutations - which now run clean under ASan and UBSan with no hangs.

+++ src/libchdr_chd.c
@@ -760,7 +760,14 @@ static CHDR_INLINE void put_bigendian_uint48(uint8_t *base, uint64_t value)
 
 static CHDR_INLINE uint32_t get_bigendian_uint32_t(const uint8_t *base)
 {
-	return (base[0] << 24) | (base[1] << 16) | (base[2] << 8) | base[3];
+	/* Cast before shifting: base[0] promotes to int, so base[0] << 24 is
+	 * signed overflow - undefined - for any byte >= 0x80. Real files never hit
+	 * it because every field read through here holds either a small count or a
+	 * four-character codec tag, and those are ASCII, but a malformed file only
+	 * has to set one high bit. The uint48 and uint64 readers above already
+	 * cast for the same reason. */
+	return ((uint32_t)base[0] << 24) | ((uint32_t)base[1] << 16) |
+	       ((uint32_t)base[2] << 8)  |  (uint32_t)base[3];
 }
 
 /*-------------------------------------------------
